Move Quadrant to own file, tweak some names
